Moved timing from misc to config
