+++ internal/config/timing.go
@@ -1,4 +1,4 @@
-package misc
+package config
 
 import (
 	"gopkg.in/yaml.v3"
